php artisan install api

+++ app/Http/Controllers/WishController.php
@@ -7,6 +7,7 @@ use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Redirect;
 use Illuminate\Support\Facades\URL;
+use Illuminate\Support\Facades\Auth;
 
 class WishController extends Controller
 {
@@ -16,14 +17,18 @@ class WishController extends Controller
 
     public function add($product_id)
     {
-        $user_id = User::first()->id;
+        if(!Auth::check()){
+            return redirect()->route('profile.edit')->with('warning', 'Veillez vous connecter pour utiliser le panier !');
+        $user_id = User::Auth()->id;
         Wish::create(['product_id' => $product_id, 'user_id' => $user_id]);
         return Redirect::to(URL::previous()."#product_$product_id")->with(['wishAdded', $product_id]);
     }
 
     public function remove($product_id)
     {
-        $user_id = User::first()->id;
+        if(!Auth::check()){
+            return redirect()->route('profile.edit')->with('warning', 'Veillez vous connecter pour utiliser le panier !');
+        $user_id = User::Auth()->id;
         $wish =  Wish::where('product_id', "=", $product_id)->where('user_id', '=', $user_id)->first();
         $wish->deleteOrFail();
         return Redirect::to(URL::previous()."#product_$product_id")->with(['wishRemoved', $product_id]);
